Permissions added

+++ app_pages/participants/participants_detail.py
@@ -26,12 +26,13 @@ class ParticipantDetails():
                 st_description = st.text_area('Description', key=f'course_details_{df_id}_description', value=f'{self.df['description'].tolist()[0]}')
                 st_active = st.toggle(label='Active', value=df['active'].tolist()[0], key=f'course_details_{df_id}_active')
                 buttons_area = st.container(horizontal=True)
-                with buttons_area:
-                    if st.button(label='', icon=':material/check:', key='participant_details_update'):
-                        self.update(name=st_name, description=st_description, active=st_active)
-                    if st.button(label='', icon=':material/delete:', key='participant_details_delete', disabled=True):
-                        self.delete()                        
-    
+                if st.user.email == 'barendkruger@gmail.com':
+                    with buttons_area:
+                        if st.button(label='', icon=':material/check:', key='participant_details_update'):
+                            self.update(name=st_name, description=st_description, active=st_active)
+                        if st.button(label='', icon=':material/delete:', key='participant_details_delete', disabled=True):
+                            self.delete()                        
+        
     @st.dialog(title='Update confirmation')
     def update(self, name=None, description=None, active=None):
         if self.df is not None:
@@ -98,8 +99,7 @@ class ParticipantDetails():
                                 if not match_participants_df.empty:
                                     for match_participant_id in match_participants_df['id'].tolist():
                                         match_participants_sql.update(id=match_participant_id, fields=fields, values=values)
-    
-    
+     
     @st.dialog(title='Delete confirmation')        
     def delete(self):
         st.write("Are you sure you want to delete the entry?")
